Add trpc endpoint for weather overlay
