+++ apps/website/src/components/overlay/Weather.tsx
@@ -1,37 +1,23 @@
-import { useEffect, useRef, useState } from "react";
+import { keepPreviousData } from "@tanstack/react-query";
 
-import { type WeatherResponse } from "@/app/api/stream/weather/route";
+import { trpc } from "@/utils/trpc";
 
 const Weather = () => {
-  // Get the current weather
-  // Refresh every 60s
-  const [weather, setWeather] = useState<WeatherResponse>();
-  const interval = useRef<NodeJS.Timeout>(null);
-  useEffect(() => {
-    const refresh = async () => {
-      const response = await fetch("/api/stream/weather");
-      if (response.ok) {
-        const data = await response.json();
-        setWeather(data);
-      }
-    };
-
-    refresh();
-    interval.current = setInterval(refresh, 60 * 1000);
-    return () => clearInterval(interval.current ?? undefined);
-  }, []);
-
-  // Remove the weather data if older than 10m
-  const stale = useRef<NodeJS.Timeout>(null);
-  useEffect(() => {
-    if (!weather) return;
-
-    stale.current = setTimeout(() => setWeather(undefined), 10 * 60 * 1000);
-    return () => clearTimeout(stale.current ?? undefined);
-  }, [weather]);
+  // Get the current weather every 60s
+  // Consider the data stale after 10 minutes
+  const { data: weather, isStale } = trpc.stream.getWeather.useQuery(
+    undefined,
+    {
+      placeholderData: keepPreviousData,
+      refetchInterval: 60 * 1000,
+      refetchIntervalInBackground: true,
+      staleTime: 10 * 60 * 1000,
+    },
+  );
 
   return (
-    weather && (
+    weather &&
+    !isStale && (
       <p className="text-3xl">
         {weather.temperature.fahrenheit}°F{" "}
         <span className="text-xl">({weather.temperature.celsius}°C)</span>
