+++ apps/website/src/server/trpc/router/_app.ts
@@ -16,6 +16,7 @@ import { formsRouter } from "./forms";
 import { notificationsRouter } from "./notifications";
 import { pushSubscriptionRouter } from "./push/subscription";
 import { showAndTellRouter } from "./show-and-tell";
+import { streamRouter } from "./stream";
 import { virtualTicketsRouter } from "./virtual-tickets";
 
 export const appRouter = router({
@@ -35,6 +36,7 @@ export const appRouter = router({
   notifications: notificationsRouter,
   pushSubscription: pushSubscriptionRouter,
   showAndTell: showAndTellRouter,
+  stream: streamRouter,
   virtualTickets: virtualTicketsRouter,
 });
