+++ apps/website/src/server/trpc/router/stream.ts
@@ -1,0 +1,6 @@
+import { publicProcedure, router } from "@/server/trpc/trpc";
+import { getWeather } from "@/server/apis/weather";
+
+export const streamRouter = router({
+  getWeather: publicProcedure.query(getWeather),
+});
